refactor: extract useUsdInspector hook from App.tsx
